+++ apps/web/src/app/project/(contants)/index.ts
@@ -1,0 +1,36 @@
+export const imagechipoptions = [
+  { value: '/images/chipIcon/javascript.svg', label: 'qz6b1owj3uqn9r5' },
+  { value: '/images/chipIcon/django.svg', label: 'no35jb96c0a99fu' },
+  { value: '/images/chipIcon/figma.svg', label: 'zzdgc96zp62tq72' },
+  { value: '/images/chipIcon/flutter.svg', label: 'ybxro0obp96yi9x' },
+  { value: '/images/chipIcon/git.svg', label: 'vhz6lpjibg6ndx5' },
+  { value: '/images/chipIcon/java.svg', label: 'hsjq2afn492t44f' },
+  { value: '/images/chipIcon/kotlin.svg', label: 'gg48yhn1z33idm5' },
+  { value: '/images/chipIcon/mongodb.svg', label: 'nxffwhetf5t3c3e' },
+  { value: '/images/chipIcon/mysql.svg', label: 'ibzw6kigzgvyqyg' },
+  { value: '/images/chipIcon/nestjs.svg', label: '0yy6wcdiqqna9tr' },
+  { value: '/images/chipIcon/nextjs.svg', label: '418yr32qvzzrg3t' },
+  { value: '/images/chipIcon/nodejs.svg', label: '6swuug4hf2tu6eo' },
+  { value: '/images/chipIcon/react.svg', label: '8cdighbw042w21w' },
+  { value: '/images/chipIcon/reactnative.svg', label: 'm80bxz9ezgr7wlm' },
+  { value: '/images/chipIcon/spring.svg', label: 'rdns4hj6q1gcno8' },
+  { value: '/images/chipIcon/swift.svg', label: 'pepzp0p6ab9vu7u' },
+  { value: '/images/chipIcon/typescript.svg', label: 'qpw5ll9u7hikykd' },
+  { value: '/images/chipIcon/vue.svg', label: 'h7wvfrlbtc7xjq7' },
+  { value: '', label: 'yel9r3y1shjyg8q' },
+];
+
+export const process = [
+  { label: '온라인', value: 'xgo777ttkdayqfl' },
+  { label: '오프라인', value: '81s1l9sgxgg9z2r' },
+  { label: '온라인/오프라인', value: '4lcuq9joet135hk' },
+];
+
+export const position = [
+  { label: '프론트엔드', value: '9f9exe2kmzcoxe6' },
+  { label: '백엔드', value: 'p8wp6pnahmb6wvj' },
+  { label: '디자이너', value: 'zk0e140xr6ly0p1' },
+  { label: 'IOS', value: 'qhxnu0nk8bvakze' },
+  { label: 'Android', value: 'es4fkkoxcmz7zoi' },
+  { label: '데브옵스', value: 's0c856g8pymem77' },
+];
