+++ apps/web/src/types/user.ts
@@ -2,6 +2,8 @@ export interface User {
   id: string;
   username: string;
   nickname: string;
+  password?: string;
+  passwordConfirm?: string;
   email: string;
   profile_image?: FormDataEntryValue | null;
   description?: string;
