+++ packages/sfac-design-kit/src/components/Chip/Chip.stories.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import { useState } from 'react';
 import { Chip, IconChip } from '.';
 
 export default {
